changed all int to int64

+++ internals/store/chat_member_store.go
@@ -15,8 +15,8 @@ const (
 )
 
 type ChatMember struct {
-	ChatID int `json:"chat_id"`
-	UserID int `json:"user_id"`
+	ChatID int64 `json:"chat_id"`
+	UserID int64 `json:"user_id"`
 	Role ChatGroupRole `json:"role"`
 	CreatedAt time.Time `json:"created_at"`
 }
@@ -32,40 +32,40 @@ func NewPostgresChatMemberStore(db *sql.DB) *PostgresChatMemberStore{
 }
 
 type ChatMemberStore interface {
-	AddMember(chatID, userID int, role string) error
-    RemoveMember(chatID, userID int) error
-    GetChatMembers(chatID int) ([]*ChatMember, error)
-    GetUserRole(chatID, userID int) (string, error)
-    IsMember(chatID, userID int) (bool, error)
-    UpdateLastRead(chatID, userID, messageID int) error
+	AddMember(chatID, userID int64, role string) error
+    RemoveMember(chatID, userID int64) error
+    GetChatMembers(chatID int64) ([]*ChatMember, error)
+    GetUserRole(chatID, userID int64) (string, error)
+    IsMember(chatID, userID int64) (bool, error)
+    UpdateLastRead(chatID, userID, messageID int64) error
 }
 
-func (pg *PostgresChatMemberStore) AddMember(chatID, userID int, role string) error {
+func (pg *PostgresChatMemberStore) AddMember(chatID, userID int64, role string) error {
 	fmt.Println("Added new member")
 	return nil
 }
 
-func (pg *PostgresChatMemberStore) RemoveMember(chatID, userID int) error {
+func (pg *PostgresChatMemberStore) RemoveMember(chatID, userID int64) error {
 	fmt.Println("removed member")
 	return nil
 }
 
-func (pg *PostgresChatMemberStore) GetChatMembers(chatID int) ([]*ChatMember, error){
+func (pg *PostgresChatMemberStore) GetChatMembers(chatID int64) ([]*ChatMember, error){
 	fmt.Println("Fetching chat members")
 	return nil, nil
 }
 
-func (pg *PostgresChatMemberStore) GetUserRole(chatID, userID int) (string, error) {
+func (pg *PostgresChatMemberStore) GetUserRole(chatID, userID int64) (string, error) {
 	fmt.Println("Fetching user role")
 	return "member", nil
 }
 
-func (pg *PostgresChatMemberStore) IsMember(chatID, userID int) (bool, error) {
+func (pg *PostgresChatMemberStore) IsMember(chatID, userID int64) (bool, error) {
 	fmt.Println("checking if this user is part of this chat")
 	return false, nil
 }
 
-func (pg *PostgresChatMemberStore) UpdateLastRead(chatID, userID, messageID int) error {
+func (pg *PostgresChatMemberStore) UpdateLastRead(chatID, userID, messageID int64) error {
 	fmt.Println("updating latest info")
 	return nil
 }
